src/third_party/blink/web_tests/virtual/streams-native/http/tests/streams/chromium/err-invalid-hwm-expected.txt

This is a testharness.js-based test.
FAIL ReadableStream RangeError on invalid highWaterMark should have an appropriate message assert_equals: message should be relevant expected "A queueing strategy's highWaterMark property must be a nonnegative, non-NaN number" but got "Failed to construct 'ReadableStream': A queuing strategy's highWaterMark property must be a nonnegative, non-NaN number"
FAIL WritableStream RangeError on invalid highWaterMark should have an appropriate message assert_equals: message should be relevant expected "A queueing strategy's highWaterMark property must be a nonnegative, non-NaN number" but got "Failed to construct 'WritableStream': A queuing strategy's highWaterMark property must be a nonnegative, non-NaN number"
Harness: the test ran to completion.
